refactor : ApiResponse 인터페이스 정의
